requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
amply==0.1.6
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
docutils==0.21.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
lxml==5.3.0
multidict==6.1.0
numpy==2.1.1
oauthlib==3.2.2
openai==0.28.1
pandas==2.2.3
PuLP==2.4
pydantic==2.9.2
pydantic_core==2.23.4
pydfs-lineup-optimizer==3.6.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.5
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1